review - rename IcebergBride.java + remove unnecessary code
